[Fix] 전체적으로 자잘한 오류 수정

+++ src/main/java/com/shinhan/soloplay/theme/ThemeEntity.java
@@ -30,7 +30,7 @@ public class ThemeEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "THEME_ID")
-    private Integer themeId;
+    private Long themeId;
 
     @ManyToOne
     @JoinColumn(name = "USER_ID")
